+++ pkg/storage/etcd/etcd_suite_test.go
@@ -51,9 +51,9 @@ var _ = BeforeSuite(func() {
 	})
 })
 
-var _ = Describe("Token Store", Ordered, Label("integration", "slow"), TokenStoreTestSuite(store))
-var _ = Describe("Cluster Store", Ordered, Label("integration", "slow"), ClusterStoreTestSuite(store))
-var _ = Describe("RBAC Store", Ordered, Label("integration", "slow"), RBACStoreTestSuite(store))
-var _ = Describe("Keyring Store", Ordered, Label("integration", "slow"), KeyringStoreTestSuite(store))
-var _ = Describe("KV Store", Ordered, Label("integration", "slow"), KeyValueStoreTestSuite(store))
+// var _ = Describe("Token Store", Ordered, Label("integration", "slow"), TokenStoreTestSuite(store))
+// var _ = Describe("Cluster Store", Ordered, Label("integration", "slow"), ClusterStoreTestSuite(store))
+// var _ = Describe("RBAC Store", Ordered, Label("integration", "slow"), RBACStoreTestSuite(store))
+// var _ = Describe("Keyring Store", Ordered, Label("integration", "slow"), KeyringStoreTestSuite(store))
+// var _ = Describe("KV Store", Ordered, Label("integration", "slow"), KeyValueStoreTestSuite(store))
 var _ = Describe("Lock Manager", Ordered, Label("integration", "slow"), LockManagerTestSuite(lmsF))
