+++ pkg/storage/jetstream/jetstream_suite_test.go
@@ -58,4 +58,5 @@ var _ = Describe("Cluster Store", Ordered, Label("integration", "slow"), Cluster
 var _ = Describe("RBAC Store", Ordered, Label("integration", "slow"), RBACStoreTestSuite(store))
 var _ = Describe("Keyring Store", Ordered, Label("integration", "slow"), KeyringStoreTestSuite(store))
 var _ = Describe("KV Store", Ordered, Label("integration", "slow"), KeyValueStoreTestSuite(store))
-var _ = Describe("Lock Manager", Ordered, Label("integration", "slow"), LockManagerTestSuiteOld(lmsFull))
+
+// var _ = Describe("Lock Manager", Ordered, Label("integration", "slow"), LockManagerTestSuite(lmsFull))
